delete stdio library in files

+++ libft.h
@@ -61,7 +61,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size);
 size_t	ft_strlcat(char *dst, const char *src, size_t size);
 
 // BONUS
-int		ft_lstsize(t_list *lst);
+size_t	ft_lstsize(t_list *lst);
 
 t_list	*ft_lstlast(t_list *lst);
 t_list	*ft_lstnew(void *content);
